Split Client & Response classes into separate files
