+++ lib/sambal/response.rb
@@ -1,0 +1,25 @@
+module Sambal
+  class Response
+
+    attr_reader :message
+
+    def initialize(message, success)
+      msg = message.split("\n")
+      msg.each do |line|
+        if line =~ /^NT\_.*\s/
+          @message = line
+        end
+      end
+      @message ||= message
+      @success = success
+    end
+
+    def success?
+      @success
+    end
+
+    def failure?
+      !success?
+    end
+  end
+end
